Update bindings to bnd 0.0.7
Migrate RDMA generation to the direct Clang, RDL, and bnd-bindgen pipeline. Regenerate the bindings, adopt the new bnd-linux module layout, and pin the matching Linux metadata artifact.

+++ rdma-io/src/pd.rs
@@ -74,7 +74,7 @@ impl ProtectionDomain {
             ibv_reg_mr(
                 self.inner,
                 buf.as_mut_ptr().cast(),
-                buf.len() as u64,
+                buf.len(),
                 access.bits() as i32,
             )
         })?;
@@ -98,7 +98,7 @@ impl ProtectionDomain {
             ibv_reg_mr(
                 self.inner,
                 buf.as_mut_ptr().cast(),
-                buf.len() as u64,
+                buf.len(),
                 access.bits() as i32,
             )
         })?;
